pyproject.toml
[project]
name = "openai-assistants-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.63.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
]
